Minion Game
Creating substring from the given string.
We evolved Stuart and Kevin as players in this game. They have to  create substring from a given string as the input. For the example given the word "Melon", the players have to produce substrings as they can from this word.

We denote Stuart to begin from the consonants while Kevin from the vowels. The winner is who can create substrings from the given word more than other. If the values show an equation means the game is draw